remove bin files

+++ OnTheRoadLibrary/src/com/ontheroad/library/GPSCoordinates.java
@@ -3,10 +3,10 @@ package com.ontheroad.library;
 import com.ontheroad.library.helpers.Mercator;
 
 public class GPSCoordinates {
-	public double Lat;
-	public double Lon;
+	public float Lat;
+	public float Lon;
 	
-	public GPSCoordinates(double lat, double lon)
+	public GPSCoordinates(float lat, float lon)
 	{
 		this.Lat = lat;
 		this.Lon = lon;
